Fix balance
