+++ src/ethUtils.js
@@ -48,4 +48,17 @@ export function isHex (h: string) {
   return out
 }
 
-export { normalizeAddress, addHexPrefix, bufToHex, validateObject, toHex }
+function unpadAddress (address: string): string {
+  const unpadded = bns.add('0', address, 16)
+  return unpadded
+}
+
+function padAddress (address: string): string {
+  const normalizedAddress = normalizeAddress(address)
+  const padding = 64 - normalizedAddress.length
+  const zeroString = '0000000000000000000000000000000000000000000000000000000000000000'
+  const out = '0x' + zeroString.slice(0, padding) + normalizedAddress
+  return out
+}
+
+export { normalizeAddress, addHexPrefix, bufToHex, validateObject, toHex, padAddress, unpadAddress }
